refactor(api): split teams error mapping into module
